# api/src/main/resources/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: Z-Commerce Platform API
  version: 1.0.0
  description: Z-Commerce e-commerce platform API for product and order management
servers:
  - url: http://localhost:8080/zcommerce
    description: Local development
  - url: http://13.202.157.59:8080/zcommerce
    description: AWS EC2
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
...
    OrderResponseApi:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderNumber:
          type: string
        customerId:
          type: integer
          format: int64
        customerEmail:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        shippingAddress:
          type: string
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponseApi'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrderRequestApi:
      type: object
      required:
        - customerId
        - customerEmail
        - orderItems
      properties:
        customerId:
          type: integer
          format: int64
        customerEmail:
          type: string
          format: email
        shippingAddress:
          type: string
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequestApi'
    OrderItemRequestApi:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
    OrderItemResponseApi:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product:
          $ref: '#/components/schemas/ProductInfo'
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
    ProductInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        sku:
          type: string
    OrderStatistics:
      type: object
      properties:
        totalOrders:
          type: integer
          format: int64
        totalRevenue:
          type: number
        averageOrderValue:
          type: number

    # Payment
    PaymentRequest:
      type: object
      required:
        - orderId
        - customerId
        - paymentMethod
        - cardNumber
        - cardHolderName
        - expiryDate
        - cvv
      properties:
        orderId:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        paymentMethod:
          type: string
          enum: [CREDIT_CARD, DEBIT_CARD, UPI, NET_BANKING]
        cardNumber:
          type: string
          pattern: '^[0-9]{13,19}$'
          description: Card number (13-19 digits)
        cardHolderName:
          type: string
          maxLength: 100
        expiryDate:
          type: string
          pattern: '^(0[1-9]|1[0-2])/([0-9]{2})$'
          description: Expiry date in MM/YY format
        cvv:
          type: string
          pattern: '^[0-9]{3,4}$'
          description: CVV (3-4 digits)
        description:
          type: string
          maxLength: 500
        customerEmail:
          type: string
          format: email
          maxLength: 100
        billingAddress:
          type: string
          maxLength: 200
        city:
          type: string
          maxLength: 50
        state:
          type: string
          maxLength: 50
        postalCode:
          type: string
          pattern: '^[0-9]{5,10}$'
        country:
          type: string
          maxLength: 50
    PaymentResponseApi:
      type: object
      properties:
        id:
          type: integer
          format: int64
        paymentId:
          type: string
        orderId:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        amount:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
        transactionId:
          type: string
        gatewayResponse:
          type: string
        failureReason:
          type: string
        processedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaymentStatistics:
      type: object
      properties:
        totalPayments:
          type: integer
          format: int64
        successfulPayments:
          type: integer
          format: int64
        failedPayments:
          type: integer
          format: int64
        totalAmount:
          type: number

    # Invoice
...
paths:
...
  /api/v1/orders:
    post:
      tags: [Orders]
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequestApi'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponseApi'

    get:
      tags: [Orders]
      operationId: getAllOrders
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
  /api/v1/orders/{orderId}:
    get:
      tags: [Orders]
      operationId: getOrderById
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponseApi'

    delete:
      tags: [Orders]
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/orders/customer/{customerId}:
    get:
      tags: [Orders]
      operationId: getOrdersByCustomer
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponseApi'
  /api/v1/orders/status/{status}:
    get:
      tags: [Orders]
      operationId: getOrdersByStatus
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponseApi'
  /api/v1/orders/{orderId}/status:
    patch:
      tags: [Orders]
      operationId: updateOrderStatus
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponseApi'
  /api/v1/orders/date-range:
    get:
      tags: [Orders]
      operationId: getOrdersByDateRange
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponseApi'
  /api/v1/orders/amount-range:
    get:
      tags: [Orders]
      operationId: getOrdersByAmountRange
      security:
        - bearerAuth: []
      parameters:
        - name: minAmount
          in: query
          required: true
          schema:
            type: number
        - name: maxAmount
          in: query
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponseApi'
  /api/v1/orders/statistics:
    get:
      tags: [Orders]
      operationId: getOrderStatistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatistics'
  /api/v1/orders/needing-attention:
    get:
      tags: [Orders]
      operationId: getOrdersNeedingAttention
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponseApi'

  # Customers
...
  /api/v1/payments:
    post:
      tags: [Payments]
      operationId: processPayment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseApi'

    get:
      tags: [Payments]
      operationId: getAllPayments
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
  /api/v1/payments/{paymentId}:
    get:
      tags: [Payments]
      operationId: getPaymentById
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseApi'

    delete:
      tags: [Payments]
      operationId: deletePayment
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/payments/order/{orderId}:
    get:
      tags: [Payments]
      operationId: getPaymentByOrderId
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseApi'
  /api/v1/payments/customer/{customerId}:
    get:
      tags: [Payments]
      operationId: getPaymentsByCustomer
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentResponseApi'
  /api/v1/payments/{paymentId}/refund:
    post:
      tags: [Payments]
      operationId: refundPayment
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseApi'
  /api/v1/payments/statistics:
    get:
      tags: [Payments]
      operationId: getPaymentStatistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatistics'

  # Invoices
...